Split whitelist logic into separate files and extend tests
