+++ Makefile
@@ -2,8 +2,8 @@ CC=gcc
 CFLAGS?=-Wall -Wextra -O2
 LDFLAGS?=-lpcap
 
-filter: main.c
-	$(CC) $(CFLAGS) $< -o $@ $(LDFLAGS)
+filter: filter_main.c whitelist.c whitelist.h
+	$(CC) $(CFLAGS) filter_main.c whitelist.c -o $@ $(LDFLAGS)
 
 test: CFLAGS+=--coverage -O0
 test: LDFLAGS+=--coverage
@@ -12,8 +12,8 @@ test: filter
 	coverage run -m pytest -v
 	coverage html -d pycov
 	coverage report -m
-	lcov --capture --directory . --output-file c_cov.info
-	genhtml c_cov.info --output-directory c_html > /dev/null
+	lcov --rc lcov_branch_coverage=1 --capture --directory . --output-file c_cov.info
+	genhtml --branch-coverage c_cov.info --output-directory c_html > /dev/null
 
 clean:
 	rm -f filter *.gcno *.gcda c_cov.info
